requirements.txt
####### project requirements.txt #######
#
# run it from cmd line:
# pip install -r requirements.txt

numpy

## data scraper requirements
bs4
requests
html5lib

# google spread sheet dependencies
gspread
oauth2client


# Loggly
loggly-python-handler
